% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_power.R
\name{wave_power}
\alias{wave_power}
\title{Wave Power calculation}
\usage{
wave_power(d, g, pi, h, t)
}
\arguments{
\item{d}{water density (kg / m^3)}

\item{g}{acceleration due to gravity (m / s ^ 2)}

\item{pi}{pi number (3.1416)}

\item{h}{wave height (m)}

\item{t}{wave period
#'}
}
\description{
Wave Power calculation
}
