% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_power.R
\name{wave_power}
\alias{wave_power}
\title{Wave Power Calculation}
\usage{
wave_power(density, gravity, pi, height, time)
}
\arguments{
\item{density}{water density (kg / m^3)}

\item{gravity}{acceleration due to gravity (m / s ^ 2)}

\item{pi}{pi number (3.1416)}

\item{height}{wave height (m)}

\item{time}{wave period
#'}
}
\description{
Wave Power Calculation
}
